backend/requirements.txt
# Core Framework
fastapi>=0.115.0,<0.120.0
uvicorn[standard]>=0.30.0,<0.35.0
python-multipart>=0.0.9
pydantic>=2.10.0
pydantic-settings>=2.5.0

# Async Support
httpx>=0.27.0
aiofiles>=24.1.0
sse-starlette>=2.1.0,<3.0.0

# LangChain & Agents (pinned to avoid pyroaring)
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
langchain-groq>=0.2.0,<0.3.0

# Database
supabase>=2.7.0,<2.12.0

# Embeddings - use httpx directly instead of voyageai
# voyageai has compatibility issues

# Caching
upstash-redis>=1.0.0,<1.2.0

# Web Search & Scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Document Processing
pymupdf>=1.24.0
python-docx>=1.1.0

# PDF Report Generation
reportlab>=4.2.0

# Monitoring
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Code Quality
ruff>=0.6.0
